Control de excepciones en eliminarEquipo de EquipoService

+++ src/main/java/madstodolist/service/EquipoService.java
@@ -81,6 +81,7 @@ public class EquipoService {
     @Transactional
     public void eliminarEquipo(Long equipoId){
         Equipo e = equipoRepository.findById(equipoId).orElse(null);
+        if(e == null) throw new EquipoServiceException("No existe el equipo con id " + equipoId);
         equipoRepository.delete(e);
     }
 }
